refactor gemm transposition variants into separate files to reduce memory usage while compiling
