+++ test/unittest/blas3/gemm/blas3_gemm_tc_test.cpp
@@ -1,0 +1,12 @@
+#include "blas_test.hpp"
+#include "../blas3_matrix_formats.hpp"
+
+typedef ::testing::Types<
+    blas_test_args<float, MatrixFormats<Transposed, Conjugate>>,
+    blas_test_args<double, MatrixFormats<Transposed, Conjugate>>
+    > BlasTypes;
+
+#define BlasTypes BlasTypes
+#define TestName gemm_normal_normal
+
+#include "blas3_gemm_def.hpp"
